refactored DriverScreens
