+++ mobile-app/api/api_Calls.js
@@ -2,7 +2,7 @@ import {useState, useEffect} from 'react'
 import axios from 'axios'
 import {API_BASE_URL} from '@env'; 
 
-const URL = "https://f89b-173-230-86-27.ngrok.io"
+const URL = "https://33f3-173-230-86-27.ngrok.io"
 
 export const GetTripByID = (tripId) => {
     console.info("Getting trip ID...");
